This branch now contains files modified to support 64-bit event codes. Everything in src and ctests has been modified and tested on a Core2 Intel machine. The Fortran interface has not been converted. Essentially all C tests pass. This branch was created as a holding pen for this code until we decide what to do with it.

+++ src/ctests/branches.c
@@ -32,7 +32,7 @@ int main(int argc, char **argv)
    double spread[MAXEVENTS];
    int nevents = MAXEVENTS;
    int eventset = PAPI_NULL;
-   int events[MAXEVENTS];
+   long long events[MAXEVENTS];
 
    events[0] = PAPI_BR_NTK;
    events[1] = PAPI_BR_PRC;
