+++ src/ctests/overflow_one_and_read.c
@@ -59,7 +59,7 @@ int main(int argc, char **argv)
    int EventSet;
    long long **values = NULL;
    int retval;
-   int PAPI_event;
+   long long PAPI_event;
    char event_name[PAPI_MAX_STR_LEN];
    int num_events1, mask1;
